refactor(fonts): use runtime as entrypoint
