+++ packages/astro/components/Font.astro
@@ -1,6 +1,6 @@
 ---
 import * as mod from 'virtual:astro:assets/fonts/internal';
-import { filterPreloads } from 'astro/assets/fonts/runtime';
+import { filterPreloads } from '../dist/assets/fonts/core/filter-preloads.js';
 import { AstroError, AstroErrorData } from '../dist/core/errors/index.js';
 
 // TODO: remove check when fonts are stabilized
