+++ packages/astro/src/assets/vite-plugin-assets.ts
@@ -149,8 +149,7 @@ export default function assets({ fs, settings, sync, logger }: Options): vite.Pl
 							export { inferRemoteSize } from "astro/assets/utils/inferRemoteSize.js";
 
 							export { default as Font } from "astro/components/Font.astro";
-							import * as fontsMod from 'virtual:astro:assets/fonts/internal';
-							import { createGetFontData } from "astro/assets/fonts/runtime";
+							export * from "astro/assets/fonts/runtime";
 							
 							export const viteFSConfig = ${JSON.stringify(resolvedConfig.server.fs ?? {})};
 
@@ -189,8 +188,6 @@ export default function assets({ fs, settings, sync, logger }: Options): vite.Pl
 								settings.config.build.assets,
 							)}, outDir);
 							export const getImage = async (options) => await getImageInternal(options, imageConfig);
-
-							export const getFontData = createGetFontData(fontsMod);
 						`,
 					};
 				}
