+++ packages/astro/test/units/assets/fonts/core.test.js
@@ -1,6 +1,8 @@
 // @ts-check
 import assert from 'node:assert/strict';
 import { describe, it } from 'node:test';
+import { createGetFontData } from '../../../../dist/assets/fonts/core/create-get-font-data.js';
+import { filterPreloads } from '../../../../dist/assets/fonts/core/filter-preloads.js';
 import { normalizeRemoteFontFaces } from '../../../../dist/assets/fonts/core/normalize-remote-font-faces.js';
 import { optimizeFallbacks } from '../../../../dist/assets/fonts/core/optimize-fallbacks.js';
 import { resolveFamily } from '../../../../dist/assets/fonts/core/resolve-families.js';
@@ -591,6 +593,176 @@ describe('fonts core', () => {
 					},
 				},
 			]);
+		});
+	});
+
+	describe('filterPreloads()', () => {
+		it('returns null if it should not preload', () => {
+			assert.equal(filterPreloads([], false), null);
+		});
+
+		it('returns everything if it should preload all', () => {
+			assert.deepStrictEqual(
+				filterPreloads(
+					[
+						{
+							style: 'normal',
+							subset: undefined,
+							type: 'woff2',
+							url: 'foo',
+							weight: undefined,
+						},
+						{
+							style: 'italic',
+							subset: 'latin',
+							type: 'otf',
+							url: 'bar',
+							weight: undefined,
+						},
+					],
+					true,
+				),
+				[
+					{
+						style: 'normal',
+						subset: undefined,
+						type: 'woff2',
+						url: 'foo',
+						weight: undefined,
+					},
+					{
+						style: 'italic',
+						subset: 'latin',
+						type: 'otf',
+						url: 'bar',
+						weight: undefined,
+					},
+				],
+			);
+		});
+
+		it('returns filtered data', () => {
+			assert.deepStrictEqual(
+				filterPreloads(
+					[
+						{
+							style: 'normal',
+							subset: undefined,
+							type: 'woff2',
+							url: 'foo',
+							weight: undefined,
+						},
+						{
+							style: 'italic',
+							subset: 'latin',
+							type: 'otf',
+							url: 'bar',
+							weight: undefined,
+						},
+					],
+					[
+						{
+							style: 'normal',
+						},
+					],
+				),
+				[
+					{
+						style: 'normal',
+						subset: undefined,
+						type: 'woff2',
+						url: 'foo',
+						weight: undefined,
+					},
+				],
+			);
+		});
+
+		it('returns variable weight', () => {
+			assert.deepStrictEqual(
+				filterPreloads(
+					[
+						{
+							style: 'normal',
+							subset: undefined,
+							type: 'woff2',
+							url: 'foo',
+							weight: '500 900',
+						},
+						{
+							style: 'italic',
+							subset: 'latin',
+							type: 'otf',
+							url: 'bar',
+							weight: '100 900',
+						},
+					],
+					[
+						{
+							weight: '400',
+						},
+					],
+				),
+				[
+					{
+						style: 'italic',
+						subset: 'latin',
+						type: 'otf',
+						url: 'bar',
+						weight: '100 900',
+					},
+				],
+			);
+
+			assert.deepStrictEqual(
+				filterPreloads(
+					[
+						{
+							style: 'normal',
+							subset: undefined,
+							type: 'woff2',
+							url: 'foo',
+							weight: '500 900',
+						},
+						{
+							style: 'italic',
+							subset: 'latin',
+							type: 'otf',
+							url: 'bar',
+							weight: '100 900',
+						},
+					],
+					[
+						{
+							weight: ' 100 900',
+						},
+					],
+				),
+				[
+					{
+						style: 'italic',
+						subset: 'latin',
+						type: 'otf',
+						url: 'bar',
+						weight: '100 900',
+					},
+				],
+			);
+		});
+	});
+
+	describe('createGetFontData()', () => {
+		it('throws if there is no consumable map', () => {
+			assert.throws(() => createGetFontData({ consumableMap: undefined })('foo'));
+		});
+
+		it('throws if no data can be found', () => {
+			assert.throws(() => createGetFontData({ consumableMap: new Map([['bar', []]]) })('foo'));
+		});
+
+		it('works when there is data', () => {
+			const data = createGetFontData({ consumableMap: new Map([['bar', []]]) })('bar');
+			assert.deepStrictEqual(data, []);
 		});
 	});
 });
